pyproject.toml
[project]
name = "nightcore-auth-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.3",
    "fastapi[standard]>=0.135.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-jose>=3.5.0",
    "redis>=7.3.0",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = ["ruff>=0.7.0"]


# ================================= RUFF
[tool.ruff]
# Line length according to PEP 8 - 79 characters
line-length = 79

# Target Python version
target-version = "py313"

exclude = [".venv", "venv", "__pycache__", "migrations", "build", "dist"]

[tool.ruff.lint]
# Enable rules
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "N",   # pep8-naming
    "D",   # pydocstyle
    "UP",  # pyupgrade
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "RUF", # Ruff-specific rules
]

# Ignore rules
ignore = ["D202", "D212", "D104", "D101", "D107", "RUF001", "RUF006"]

# Allow unused variables if they start with an underscore
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Docstring handling
[tool.ruff.lint.pydocstyle]
convention = "google"

# Settings for isort
[tool.ruff.lint.isort]
known-first-party = ["src"]
section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder",
]

# Formatting options
[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.pyright]
include = ["src"]
pythonVersion = "3.13"
typeCheckingMode = "strict"
